Introduce `sc-network-bitswap`
Move the bitswap protocol out of `sc-network` to its own crate.
