+++ client/network/Cargo.toml
@@ -13,9 +13,6 @@ readme = "README.md"
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
 
-[build-dependencies]
-prost-build = "0.10"
-
 [dependencies]
 async-trait = "0.1"
 asynchronous-codec = "0.6"
@@ -50,6 +47,7 @@ prometheus-endpoint = { package = "substrate-prometheus-endpoint", version = "0.
 sc-block-builder = { version = "0.10.0-dev", path = "../block-builder" }
 sc-client-api = { version = "4.0.0-dev", path = "../api" }
 sc-consensus = { version = "0.10.0-dev", path = "../consensus/common" }
+sc-network-bitswap = { version = "0.10.0-dev", path = "./bitswap" }
 sc-network-common = { version = "0.10.0-dev", path = "./common" }
 sc-peerset = { version = "4.0.0-dev", path = "../peerset" }
 sc-utils = { version = "4.0.0-dev", path = "../utils" }
