+++ client/network/bitswap/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+description = "Substrate transaction protocol"
+name = "sc-network-bitswap"
+version = "0.10.0-dev"
+license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2021"
+homepage = "https://substrate.io"
+repository = "https://github.com/paritytech/substrate/"
+documentation = "https://docs.rs/sc-network-bitswap"
+readme = "README.md"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[build-dependencies]
+prost-build = "0.10"
+
+[dependencies]
+cid = "0.8.4"
+futures = "0.3.21"
+libp2p = "0.46.1"
+log = "0.4.17"
+prost = "0.10"
+thiserror = "1.0"
+unsigned-varint = { version = "0.7.1", features = ["futures", "asynchronous_codec"] }
+void = "1.0.2"
+sc-client-api = { version = "4.0.0-dev", path = "../../api" }
+sc-network-common = { version = "0.10.0-dev", path = "../common" }
+sp-blockchain = { version = "4.0.0-dev", path = "../../../primitives/blockchain" }
+sp-runtime = { version = "6.0.0", path = "../../../primitives/runtime" }
