+++ client/network/bitswap/build.rs
@@ -1,4 +1,4 @@
-const PROTOS: &[&str] = &["src/schema/bitswap.v1.2.0.proto"];
+const PROTOS: &[&str] = &["bitswap.v1.2.0.proto"];
 
 fn main() {
 	prost_build::compile_protos(PROTOS, &["src/schema"]).unwrap();
