+++ client/network/src/lib.rs
@@ -249,12 +249,10 @@ mod discovery;
 mod peer_info;
 mod protocol;
 mod request_responses;
-mod schema;
 mod service;
 mod transport;
 mod utils;
 
-pub mod bitswap;
 pub mod config;
 pub mod error;
 pub mod network_state;
