+++ client/network/test/src/lib.rs
@@ -883,7 +883,7 @@ where
 			import_queue,
 			chain_sync: Box::new(chain_sync),
 			metrics_registry: None,
-			bitswap: None,
+			bitswap_protocol_config: None,
 			block_request_protocol_config,
 			state_request_protocol_config,
 			light_client_request_protocol_config,
